cfdna-epigenomics-mini/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cfdna-epigenomics-mini"
version = "0.1.0"
description = "Early cancer detection from cfDNA epigenomic signals"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "cfDNA Team"}]
requires-python = ">=3.11"
dependencies = [
    "numpy",
    "scipy", 
    "pandas",
    "scikit-learn",
    "pyarrow",
    "matplotlib",
    "seaborn",
    "statsmodels",
    "torch",
    "tqdm",
    "pyyaml",
    "sqlalchemy",
    "plotnine",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "pre-commit",
    "jupyter",
    "ipykernel",
]

[project.scripts]
cfdna-sim = "cfdna.simulate:main"
cfdna-train = "cfdna.train:main"
cfdna-eval = "cfdna.eval:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings  
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = ["E501"]  # line too long

[tool.ruff.lint.isort]
known-first-party = ["cfdna"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"